Python List to Array conversion and back
We convert py List to js Array and back. Type List supports a strict subset of type Array, not an intersection. So when we go py to js we convert List to true Array, and when we go the other way we drop all of Array that is not supported by List